// src/Interfaces/Rdv.tsx
interface RDV {
    beginning: string
    end: Date | undefined
    description: string | undefined
    lastname: string | undefined
    firstname: string | undefined
    mail: string | undefined
    phone: string | undefined
    address: string | undefined
    city: string | undefined
    zipcode: string | undefined
    label: string | undefined
    name: string | undefined
}

export default RDV